# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ConflictBody": {
        "description": "409 — a ``block`` rule fired; repeat the write with ``override`` to keep it.",
        "properties": {
          "error": {
            "const": "conflict",
            "title": "Error",
            "type": "string"
          },
          "violations": {
            "items": {
              "$ref": "#/components/schemas/ViolationBody"
            },
            "title": "Violations",
            "type": "array"
          }
        },
        "required": [
          "error",
          "violations"
        ],
        "title": "ConflictBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HevySyncResponse": {
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "synced": {
            "default": 0,
            "title": "Synced",
            "type": "integer"
          }
        },
        "required": [
          "ok"
        ],
        "title": "HevySyncResponse",
        "type": "object"
      },
...
      "InvalidBody": {
        "description": "400 — a service refused the values (its ``ValueError`` message).",
        "properties": {
          "error": {
            "const": "invalid",
            "title": "Error",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "title": "InvalidBody",
        "type": "object"
      },
...
      "MacroSplit": {
        "properties": {
          "carbs_pct": {
            "title": "Carbs Pct",
            "type": "number"
          },
          "fat_pct": {
            "title": "Fat Pct",
            "type": "number"
          },
          "protein_pct": {
            "title": "Protein Pct",
            "type": "number"
          }
        },
        "required": [
          "protein_pct",
          "fat_pct",
          "carbs_pct"
        ],
        "title": "MacroSplit",
        "type": "object"
      },
...
      "MealCreate": {
        "properties": {
          "calories": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calories"
          },
          "carbs_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carbs G"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "fat_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fat G"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "override": {
            "default": false,
            "title": "Override",
            "type": "boolean"
          },
          "protein_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protein G"
          },
          "time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time"
          }
        },
        "required": [
          "date",
          "name"
        ],
        "title": "MealCreate",
        "type": "object"
      },
      "MealItem": {
        "properties": {
          "calories": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calories"
          },
          "carbs_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carbs G"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "fat_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fat G"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "protein_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protein G"
          },
          "time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time"
          }
        },
        "required": [
          "id",
          "date",
          "name"
        ],
        "title": "MealItem",
        "type": "object"
      },
      "MealPatch": {
        "properties": {
          "calories": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calories"
          },
          "carbs_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carbs G"
          },
          "date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "fat_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fat G"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "override": {
            "default": false,
            "title": "Override",
            "type": "boolean"
          },
          "protein_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protein G"
          },
          "time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time"
          }
        },
        "title": "MealPatch",
        "type": "object"
      },
      "MealRef": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id"
        ],
        "title": "MealRef",
        "type": "object"
      },
...
      "NotFoundBody": {
        "description": "404 — the module behind the endpoint is switched off, or nothing is there.",
        "properties": {
          "error": {
            "enum": [
              "module_disabled",
              "not_found"
            ],
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "NotFoundBody",
        "type": "object"
      },
...
      "NutritionGoals": {
        "properties": {
          "calories_max": {
            "title": "Calories Max",
            "type": "number"
          },
          "calories_min": {
            "title": "Calories Min",
            "type": "number"
          },
          "carbs_target_g": {
            "title": "Carbs Target G",
            "type": "number"
          },
          "fat_target_g": {
            "title": "Fat Target G",
            "type": "number"
          },
          "protein_target_g": {
            "title": "Protein Target G",
            "type": "number"
          }
        },
        "required": [
          "calories_min",
          "calories_max",
          "protein_target_g",
          "fat_target_g",
          "carbs_target_g"
        ],
        "title": "NutritionGoals",
        "type": "object"
      },
      "NutritionView": {
        "properties": {
          "calories_pct": {
            "title": "Calories Pct",
            "type": "number"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "goals": {
            "$ref": "#/components/schemas/NutritionGoals"
          },
          "is_today": {
            "title": "Is Today",
            "type": "boolean"
          },
          "macro_split": {
            "$ref": "#/components/schemas/MacroSplit"
          },
          "meals": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/MealItem"
            },
            "title": "Meals",
            "type": "array"
          },
          "next_date": {
            "format": "date",
            "title": "Next Date",
            "type": "string"
          },
          "prev_date": {
            "format": "date",
            "title": "Prev Date",
            "type": "string"
          },
          "protein_pct": {
            "title": "Protein Pct",
            "type": "number"
          },
          "recent_days": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/NutritionDayMini"
            },
            "title": "Recent Days",
            "type": "array"
          },
          "today_date": {
            "format": "date",
            "title": "Today Date",
            "type": "string"
          },
          "totals": {
            "$ref": "#/components/schemas/MacroTotals"
          }
        },
        "required": [
          "date",
          "today_date",
          "is_today",
          "prev_date",
          "next_date",
          "totals",
          "goals",
          "calories_pct",
          "protein_pct",
          "macro_split"
        ],
        "title": "NutritionView",
        "type": "object"
      },
...
      "UnauthenticatedBody": {
        "description": "401 — no valid session cookie.",
        "properties": {
          "error": {
            "const": "unauthenticated",
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "UnauthenticatedBody",
        "type": "object"
      },
...
      "WorkoutsView": {
        "properties": {
          "catalog": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ExerciseCatalogItem"
            },
            "title": "Catalog",
            "type": "array"
          },
          "exercise_count": {
            "title": "Exercise Count",
            "type": "integer"
          },
          "is_configured": {
            "title": "Is Configured",
            "type": "boolean"
          },
          "last_sync": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync"
          },
          "last_workout_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Workout Date"
          },
          "workout_count": {
            "title": "Workout Count",
            "type": "integer"
          },
          "workouts": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/HevyWorkoutItem"
            },
            "title": "Workouts",
            "type": "array"
          }
        },
        "required": [
          "workout_count",
          "exercise_count",
          "is_configured"
        ],
        "title": "WorkoutsView",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Vitals API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/nutrition": {
      "get": {
        "description": "The Nutrition screen: daily KBJU totals against goals, macro split, meals list,\nand 30-day mini bars.",
        "operationId": "read_nutrition",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NutritionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Nutrition"
      }
    },
    "/api/v1/nutrition/meals": {
      "post": {
        "description": "Log a meal.",
        "operationId": "create_meal",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MealCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealRef"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Meal"
      }
    },
    "/api/v1/nutrition/meals/{meal_id}": {
      "delete": {
        "description": "Delete a meal entry.",
        "operationId": "delete_meal",
        "parameters": [
          {
            "in": "path",
            "name": "meal_id",
            "required": true,
            "schema": {
              "title": "Meal Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Meal"
      },
      "patch": {
        "description": "Edit a logged meal.",
        "operationId": "update_meal",
        "parameters": [
          {
            "in": "path",
            "name": "meal_id",
            "required": true,
            "schema": {
              "title": "Meal Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MealPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealRef"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Meal"
      }
    },
...
    "/api/v1/workouts": {
      "get": {
        "description": "The Workouts screen: recent sessions expandable to sets, exercise catalog with progression.",
        "operationId": "read_workouts",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutsView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundBody"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "summary": "Read Workouts"
      }
    },
    "/api/v1/workouts/sync": {
      "post": {
        "description": "Pull the latest workouts from Hevy.",
        "operationId": "sync_hevy_now",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HevySyncResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthenticatedBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundBody"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "summary": "Sync Hevy Now"
      }
    }
  }
}